Refactor models into separate apps
