+++ core/models.py
@@ -1,73 +1,9 @@
-"""
-Database models for the Quran memorization assistant.
-
-These models define the fundamental data structures required
-to support users (students), complaints, Quranic metadata, and
-test sessions. They are simplified placeholders; relationships
-can be expanded to cover more advanced features.
-"""
-from django.db import models  # type: ignore
-from django.contrib.auth.models import User  # type: ignore
-
-
-class Student(models.Model):
-    avatar = models.ImageField(upload_to='avatars/', blank=True, null=True)
-    def avatar_url(self):
-        try:
-            if self.avatar and hasattr(self.avatar, "url"):
-                return self.avatar.url
-        except Exception:
-            pass
-        return ""  # هنستخدم شكل افتراضي بالـCSS لو فاضي
-    """A simple student profile linked to the built-in User model."""
-
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    display_name = models.CharField(max_length=100)
-    skin = models.CharField(max_length=50, default='default', help_text='Skin theme for the student')
-
-    def __str__(self) -> str:
-        return self.display_name
-
-
-class Complaint(models.Model):
-    """A complaint or suggestion submitted by a student."""
-
-    student = models.ForeignKey(Student, on_delete=models.CASCADE)
-    text = models.TextField()
-    created_at = models.DateTimeField(auto_now_add=True)
-    resolved = models.BooleanField(default=False)
-
-    def __str__(self) -> str:
-        return f"Complaint by {self.student}: {self.text[:30]}"
-
-
-class Juz(models.Model):
-    """Represents one of the 30 parts of the Qur’an."""
-
-    number = models.PositiveSmallIntegerField(unique=True)
-    name = models.CharField(max_length=50, blank=True)
-
-    def __str__(self) -> str:
-        return f"Juz {self.number}"
-
-
-class Quarter(models.Model):
-    """Represents a quarter of a Juz (Rubʿ)."""
-
-    juz = models.ForeignKey(Juz, on_delete=models.CASCADE)
-    index_in_juz = models.PositiveSmallIntegerField(help_text="1–8 for each Juz")
-    label = models.CharField(max_length=100, help_text="Name of the quarter from the opening words of its first verse")
-
-    class Meta:
-        unique_together = ('juz', 'index_in_juz')
-
-    def __str__(self) -> str:
-        return f"{self.juz} - Quarter {self.index_in_juz}"
+"""Core models unrelated to the specialized apps."""
+from django.db import models
 
 
 class Page(models.Model):
     """Represents a page of the Mushaf with its image."""
-
     number = models.PositiveSmallIntegerField(unique=True)
     image = models.ImageField(upload_to='pages/')
 
@@ -75,27 +11,10 @@ class Page(models.Model):
         return f"Page {self.number}"
 
 
-class Ayah(models.Model):
-    """Represents a single ayah with its metadata."""
-
-    surah = models.PositiveSmallIntegerField()
-    number = models.PositiveSmallIntegerField()
-    text = models.TextField()
-    page = models.ForeignKey(Page, on_delete=models.SET_NULL, null=True, blank=True)
-    quarter = models.ForeignKey(Quarter, on_delete=models.SET_NULL, null=True, blank=True)
-
-    class Meta:
-        unique_together = ('surah', 'number')
-
-    def __str__(self) -> str:
-        return f"{self.surah}:{self.number}"
-
-
 class SimilarityGroup(models.Model):
     """Group of similar verses (mutashabehat)."""
-
     name = models.CharField(max_length=200, help_text="Key phrase that identifies the group")
-    ayat = models.ManyToManyField(Ayah, related_name='similarity_groups')
+    ayat = models.ManyToManyField('quran_structure.Ayah', related_name='similarity_groups')
 
     def __str__(self) -> str:
         return self.name
@@ -114,7 +33,7 @@ class Phrase(models.Model):
 
 class PhraseOccurrence(models.Model):
     phrase = models.ForeignKey(Phrase, related_name='occurrences', on_delete=models.CASCADE)
-    ayah = models.ForeignKey(Ayah, on_delete=models.CASCADE)
+    ayah = models.ForeignKey('quran_structure.Ayah', on_delete=models.CASCADE)
     start_word = models.PositiveSmallIntegerField()
     end_word = models.PositiveSmallIntegerField()
 
@@ -124,78 +43,3 @@ class PhraseOccurrence(models.Model):
             models.Index(fields=['ayah']),
             models.Index(fields=['phrase']),
         ]
-
-
-class TestSession(models.Model):
-    """Represents one attempt of a test by a student."""
-
-    # NEW: أنواع الاختبارات (الجديدة + القديمة حفاظًا على التوافق)
-    TEST_TYPE_CHOICES = [
-        ('similar_count', 'عدد مواضع المتشابهات'),                 # المتاح حاليًا
-        ('similar_on_pages', 'مواضع المتشابهات في الصفحات'),        # قريبًا
-        ('page_edges_quarters', 'بداية/نهاية الصفحات مع الأرباع'),   # قريبًا
-        ('order_juz_quarters', 'ترتيب الأجزاء والأرباع'),            # قريبًا
-        ('semantic_similarities', 'متشابهات معاني الآيات'),          # قريبًا
-        ('verse_location_quarters', 'موقع الآية في الربع والصفحة'),   # الجديد
-
-        # القيم القديمة إن كانت موجودة في بيانات سابقة:
-        ('similar_only', 'Similar verses only'),
-        ('similar_quarters', 'Similar verses & quarters'),
-        ('similar_quarters_location', 'Similar verses & quarters with location'),
-        ('mixed', 'Mixed'),
-    ]
-
-    student = models.ForeignKey(Student, on_delete=models.CASCADE)
-    created_at = models.DateTimeField(auto_now_add=True)
-    started_at = models.DateTimeField(null=True, blank=True, help_text='وقت بداية الاختبار')
-    test_type = models.CharField(
-        max_length=50,
-        choices=TEST_TYPE_CHOICES,
-        default='similar_count',
-    )
-    num_questions = models.PositiveSmallIntegerField(default=10)
-    difficulty = models.CharField(
-        max_length=10,
-        choices=[('easy', 'Easy'), ('medium', 'Medium'), ('hard', 'Hard')],
-        default='easy',
-    )
-    completed = models.BooleanField(default=False)
-    position_order = models.CharField(
-        max_length=20,
-        choices=[('normal', 'عادي'), ('reverse', 'عكسي')],
-        default='normal',
-        help_text='ترتيب المواضع في الأسئلة'
-    )
-
-    # ManyToMany to Juz and Quarter representing scope
-    juzs = models.ManyToManyField(Juz, blank=True)
-    quarters = models.ManyToManyField(Quarter, blank=True)
-    
-    # Results fields
-    completed_at = models.DateTimeField(null=True, blank=True, help_text='وقت انتهاء الاختبار')
-    total_questions = models.PositiveSmallIntegerField(null=True, blank=True, help_text='إجمالي عدد الأسئلة')
-    correct_answers = models.PositiveSmallIntegerField(null=True, blank=True, help_text='عدد الإجابات الصحيحة')
-    wrong_answers = models.PositiveSmallIntegerField(null=True, blank=True, help_text='عدد الإجابات الخاطئة')
-    unanswered = models.PositiveSmallIntegerField(null=True, blank=True, help_text='عدد الأسئلة غير المجاب عليها')
-
-    def __str__(self) -> str:
-        return f"TestSession {self.id} for {self.student}"
-
-
-class TestQuestion(models.Model):
-    """Single question within a test session."""
-
-    session = models.ForeignKey(TestSession, on_delete=models.CASCADE, related_name='questions')
-    similarity_group = models.ForeignKey(SimilarityGroup, on_delete=models.SET_NULL, null=True)
-    phrase = models.ForeignKey(Phrase, on_delete=models.SET_NULL, null=True, blank=True)
-    question_type = models.CharField(max_length=50, default='similar_count')
-    question_text = models.TextField(blank=True, null=True)
-    correct_answer = models.TextField(blank=True, null=True)
-    # store JSON or text representing the student's answer. In a complete implementation
-    # this could be structured data (quarter ID, page number, half, etc.).
-    student_response = models.TextField(blank=True)
-    is_correct = models.BooleanField(default=False)
-    answered_at = models.DateTimeField(null=True, blank=True, help_text='وقت الإجابة على السؤال')
-
-    def __str__(self) -> str:
-        return f"Question {self.id} in {self.session}"
