+++ core/signals.py
@@ -1,7 +1,7 @@
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 from django.contrib.auth.models import User
-from .models import Student
+from students.models import Student
 from .services.user_service import UserService
 
 @receiver(post_save, sender=User)
