+++ quran_helper/settings.py
@@ -33,6 +33,9 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "core",
+    "students",
+    "quran_structure",
+    "testing",
     "tests_app",
     "mushaf_app",
     "complaints_app",
