+++ test_similar_phrases.py
@@ -5,7 +5,8 @@ import math
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'quran_helper.settings')
 django.setup()
 
-from core.models import PhraseOccurrence, Phrase, Ayah
+from core.models import PhraseOccurrence, Phrase
+from quran_structure.models import Ayah
 from django.db.models import Count
 
 print("=== اختبار منطق البحث عن العبارات المتشابهة ===")
