+++ tests_app/similar_on_pages/views_cbv.py
@@ -5,7 +5,7 @@ from django.contrib import messages
 from django.urls import reverse_lazy
 from django.http import Http404
 
-from core.models import TestSession
+from testing.models import TestSession
 from tests_app.services.test_service import TestService
 from core.services.user_service import UserService
 
@@ -27,7 +27,7 @@ class SelectionView(LoginRequiredMixin, TemplateView):
         context['student'] = student
         context['selected_test_type'] = 'similar_on_pages'
 
-        from core.models import Juz, Quarter
+        from quran_structure.models import Juz, Quarter
         juz_quarters_map = {}
         for juz in Juz.objects.all():
             quarters = Quarter.objects.filter(juz=juz).order_by('index_in_juz')
